requirements.txt
Flask==2.2.3
openai==0.27.8
python-dotenv==1.0.0
gunicorn==20.1.0
Flask-Session==0.4.0
